gitserver: Move urlredactor to separate package
Self-contained code piece, so it can live in it's own package, decluttering the massive gitserver package.
