+++ cmd/gitserver/server/urlredactor/BUILD.bazel
@@ -1,0 +1,17 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("//dev:go_defs.bzl", "go_test")
+
+go_library(
+    name = "urlredactor",
+    srcs = ["urlredactor.go"],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/gitserver/server/urlredactor",
+    visibility = ["//visibility:public"],
+    deps = ["//internal/vcs"],
+)
+
+go_test(
+    name = "urlredactor_test",
+    srcs = ["urlredactor_test.go"],
+    embed = [":urlredactor"],
+    deps = ["//internal/vcs"],
+)
